Update _setup_hostname to write to /etc/hostname file when failing on hostnamectl. Move function to base class since function is the smae for all operating systems
